# Add multiple files compilation mode for crossgen2