By using `--out-near-input` and `--single-file-compilation` options multiple files can be compiled in one invocation of crossgen2. This allows to remove startup overhead if many files are to be compiled anyway.

x64, f95b2b2fa2d290a2a79e21677a1f326284845ecb, release build, 100 measurements for each case (for single-file compilation mode all 100 copies of file are passed in one command):
```sh
./corerun `pwd`/crossgen2/crossgen2.dll /tmp/crossgen2.dll -O -r:`pwd`/*  --out-near-input --single-file-compilation
```

### default number of threads

Basically, the result is a constant ~0.3s diff per dll:
```
before:
crossgen2.dll: 0.80555s
System.Private.CoreLib.dll: 3.15564s

after:
crossgen2.dll: 0.54847s (-31.9%, -0.26s)
System.Private.CoreLib.dll: 2.80951s (-11.0%, -0.35s) 
```

### 1 thread
```
before:
crossgen2.dll: 0.88498s
System.Private.CoreLib.dll: 8.32537s

after:
crossgen2.dll: 0.53211s (-39.9%, -0.35s)
System.Private.CoreLib.dll: 7.74492s (-7.0%, -0.58s) 
```


cc @alpencolt