@MichalStrehovsky 2020-06-04
Did you run crossgen on crossgen2 itself before doing the measurements? Without R2R code, a significant portion of time will be spent JITting the compiler.

Compilers (clang, Roslyn, VC++, you name it) don't offer multiple input/multiple output modes because they're not useful when integrating into build systems. Multiple-input/multiple-output is a test hook. We place test hooks in the src\coreclr\src\tools\r2rtest runner so that we don't have test hooks in the shipping compiler. r2rtest already has modes to compile all files in a directory.

We could potentially add a new launch option that doesn't create a new compilation process, but does an `Assembly.Load` to load the crossgen2 entrypoint assembly if this is really important. Once the assembly is loaded it can do `asm.EntryPoint.Invoke(...)` to invoke the entrypoint in the context of the current process.

Note that the crossgen2 compiler has known "memory leaks" when run in such mode and will run out of memory eventually (there are static fields that cache state that is only relevant to a single compilation and is never released until the process dies).

Cc @dotnet/crossgen-contrib