@gbalykov 2020-06-04
Yes, forgot to mention that, all system libs and crossgen2 libs are compiled with first crossgen in r2r mode. However, runtime still spends much time on jitting, even with r2r images. Change in this PR allows to mitigate startup overhead of jitting and loading crossgen2.

Our concern is that on small dlls crossgen2 perf is much worse than first crossgen (for example, on more that 1100% for crossgen2.dll on x64). For large dlls like SPC.dll crossgen2 is better on 25% on x64, however this is achieved with 16 threads, and with 1 thread crossgen2 is 2x times slower. 

On arm devices with 2 cpus, however, we can't launch so many threads, so crossgen2 is slower than first crossgen even on SPC.dll.

Here's some data:
#### x64

|dll|threads|crossgen, s|crossgen2, s|diff(%)|
|---|---|---|---|---|
|crossgen2.dll|default|0.06508|0.80555|+1138|
| |1|0.06508|0.88498|+1260|
| |2| |0.85827|+1219|
| |4| |0.81757|+1172|
| |8| |0.8094|+1144|
| |16| |0.80782|+1141|
| |32| |0.80494|+1137|
|System.Private.CoreLib.dll|default|4.18926|3.15564| -24.7|
| |1|4.18926|8.32537|+98.7|
| |2| |5.55454|+32.6|
| |4| |4.14622|-1.0|
| |8| |3.3941|-19.0|
| |16| |3.12814|-25.3|
| |32| |3.17038|-24.3|

#### armel

|dll|threads|crossgen|crossgen2|diff(%)|
|---|---|---|---|---|
|crossgen2.dll|default|0.56884|7.10501|+1149|
| |1|0.56884|7.21069|+1167|
| |2| |6.94721|+1121|
| |4| |6.88047|+1110|
| |8| |6.80611|+1096|
|System.Private.CoreLib.dll|default|43.1077| | |
| |1|43.1077|88.631|+105.6|
| |2| |61.0679|+41.7|
| |4| |61.4447|+42.5|
| |8| |61.9229|+43.6|