@davidwrighton 2020-06-04
@gbalykov, thank you for looking deeper into this. I agree that it is quite concerning how slow crossgen2 is, especially for the smaller binaries. There are several details to note.

1. Performance of the compiler running on the armel platform is very slow, and I assume that is significant in terms of engineering cost for your team. Have you considered building a cross-targetting jit so that you can run the armel compiles on an x64 machine, or does that not fit with your engineering strategy? Based on the numbers from your most recent post here, that would be a better time improvement than making this change to the crossgen2 compilation model? We have done some experiments with an arm64 targetting cross compiler and found that we were easily able to produce binary identical images while running on an x64 host. 

2. @MichalStrehovsky This change in compilation model does have precedent in the .NET ecosystem. In particular, Roslyn actually is typically run in a mode which uses a compiler server which amortizes the cost of jitting the Roslyn codebase across multiple compilations. In that mode, there is a compiler server which does the actual compilations, and csc just serves to pass command line arguments to that persistent server. That preserves the illusion that the compiler has a single output model to the build system, but also allows for multiple processes to take advantage of the benefits of running the compiler without needing excess jitting. Of course, this sort of server model requires some careful engineering to make sure that repeated compilations do not interfere with each others, etc.

3. We expect that the composite build mode of crossgen2 may provide benefits for this sort of situation, such that a R2R variant of crossgen2 may be closer in performance to crossgen. We don't have good numbers for that yet, but in the next couple of months we hope to get them.

4. Given the above, and especially point 2, I'm tempted to suggest that if you are unable to use a cross compilation approach, instead of building in new command line options, we would be more likely to accept a patch, which moved much of this processing in crossgen2.dll into ILCompiler.ReadyToRun.dll, such that it could be run by either crossgen2 as an application, or used directly by some sort of wrapper or server process that would be able to achieve these sorts of performance wins. I think that's a bit more work, but long term, I believe the compiler server approach would be better than a series of new command line options as is proposed here.