@MichalStrehovsky 2020-06-05
> This change in compilation model does have precedent in the .NET ecosystem. In particular, Roslyn actually is typically run in a mode which uses a compiler server which amortizes the cost of jitting the Roslyn codebase across multiple compilations.

Yes, I'm aware of that mode. That's why I keep pushing for people to stop using statics to store per compilation state but people [keep adding those](https://github.com/dotnet/runtime/pull/34842/files#diff-88edef5be22c6bc8de64aa79a0cc901aR112) when I'm not looking. But there's a difference between a compiler server and a command line argument to batch compile. The former can be integrated into build systems (but it does bring it's own challenges which is why I sometimes need to taskkill `dotnet` process before I can do `git clean` on the runtime repo); the latter cannot be integrated into build systems and cannot be a shipping switch.

I know what sort of response I get for this, but I compiled crossgen2 with the CoreRT compiler and compared throughput with the CoreCLR/ReadyToRun based one:

|                                | Before  | After   | Improvement |
|--------------------------------|---------|---------|-------------|
| Compile System.Private.CoreLib | 3446 ms | 2632 ms | 23%         |
| Compile Hello World            | 452 ms  | 68 ms   | 85%         |

We'll probably need to build a compilation server to get throughput anywhere near this as long as the compiler is hosted on top of CoreCLR.