@MichalStrehovsky 2020-06-05
> We expect that the composite build mode of crossgen2 may provide benefits for this sort of situation

When we were discussing publishing options for crossgen2, we ruled out self-contained publishing because the size was prohibitively large - I don't think we'll be able to composite-compile crossgen2 itself and ship it that way to our customers. We can use it to speed up our inner loop, but I'm skeptical of our ability to pass the benefit to our customers.